pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "multi-agent-ops-assistant"
version = "1.0.0"
description = "Multi-agent operations assistant using Microsoft Agent Framework + Foundry"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # ── Microsoft Agent Framework ──
    "agent-framework-azure-ai==1.0.0rc2",
    "agent-framework-azure-ai-search==1.0.0b260225",
    "agent-framework-core==1.0.0rc2",

    # ── Azure AI Services ──
    "azure-ai-projects==2.0.0b3",
    "azure-ai-evaluation==1.15.3",
    "azure-ai-contentsafety==1.0.0",

    # ── Azure Observability ──
    "azure-monitor-opentelemetry==1.8.5",

    # ── Azure Identity & Search ──
    "azure-identity==1.26.0b2",
    "azure-search-documents==11.7.0b2",

    # ── OpenAI ──
    "openai==2.24.0",

    # ── Web Framework ──
    "fastapi==0.134.0",
    "uvicorn[standard]==0.41.0",
    "websockets==16.0",

    # ── Data & Config ──
    "python-dotenv==1.2.1",
    "pyodbc==5.3.0",
    "pydantic==2.13.0b2",
    "pyyaml==6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
